+++ src/controllers/customer-controller.ts
@@ -1,5 +1,5 @@
 import { Request, Response } from "express";
-import { CustomerService } from "../services/customerService.js";
+import { CustomerService } from "../services/customer-service.js";
 
 const customerService = new CustomerService();
 
@@ -7,7 +7,7 @@ export class CustomerController {
   // 고객 등록
   async create(req: Request, res: Response) {
     try {
-      // ✅ Type Narrowing: req.user 존재 확인
+      //req.user 존재 확인
       if (!req.user) {
         return res.status(401).json({ message: "인증된 사용자가 아닙니다." });
       }
@@ -20,28 +20,32 @@ export class CustomerController {
     }
   }
 
-  // 고객 목록 조회 (검색 + 페이지네이션)
-  async list(req: Request, res: Response) {
-    try {
-      if (!req.user) {
-        return res.status(401).json({ message: "인증된 사용자가 아닙니다." });
-      }
-
-      const companyId = req.user.companyId;
-      const { search, page, limit } = req.query;
-
-      const result = await customerService.getCustomers(
-        companyId,
-        search ? String(search) : undefined,
-        page ? Number(page) : 1,
-        limit ? Number(limit) : 10
-      );
-
-      return res.status(200).json(result);
-    } catch (error: any) {
-      return res.status(400).json({ message: error.message });
+ // 고객 목록 조회 (검색 + 페이지네이션)
+async list(req: Request, res: Response) {
+  try {
+    if (!req.user) {
+      return res.status(401).json({ message: "인증된 사용자가 아닙니다." });
     }
+
+    const companyId = req.user.companyId;
+
+    // 프론트엔드 요청에 맞게 변수명 수정
+    const { page, pageSize, searchBy, keyword } = req.query;
+
+    const result = await customerService.getCustomers(
+      companyId,
+      String(searchBy || "name"), // 검색 기준
+      String(keyword || ""),      // 검색어
+      Number(page) || 1,          // 페이지 번호
+      Number(pageSize) || 10      // 페이지당 데이터 수
+    );
+
+    return res.status(200).json(result);
+  } catch (error: any) {
+    return res.status(400).json({ message: error.message });
   }
+}
+
 
   // 고객 상세 조회
   async detail(req: Request, res: Response) {
